+++ src/main/php/rdbms/ResultSet.class.php
@@ -1,39 +1,34 @@
 <?php namespace rdbms;
 
-use util\Date;
+use util\TimeZone;
 use lang\Closeable;
-
 
 /**
  * Result set as returned from the DBConnection::query method
  *
  * Usage (abbreviated example):
- * <code>
- *   // [...]
- *   $r= $conn->query('select news_id, caption, created_at from news');
- *   while ($row= $r->next()) {
- *     var_dump($row);
- *   }
- *   $r->close();
- *   // [...]
- * </code>
+ * ```php
+ * $r= $conn->query('select news_id, caption, created_at from news');
+ * while ($row= $r->next()) {
+ *   var_dump($row);
+ * }
+ * $r->close();
+ * ```
  *
- * @purpose  Resultset wrapper
+ * @test  xp://rdbms.unittest.ResultSetTest
  */
-class ResultSet extends \lang\Object implements Closeable {
-  public
-    $handle,
-    $fields,
-    $tz;
+abstract class ResultSet extends \lang\Object implements Closeable, \IteratorAggregate {
+  protected $handle, $fields, $tz;
+  private $iterator= null;
 
   /**
    * Constructor
    *
-   * @param   resource handle
-   * @param   array fields
-   * @param   util.TimeZone tz default null
+   * @param   var $handle
+   * @param   var $fields
+   * @param   util.TimeZone $tz
    */
-  public function __construct($handle, $fields, \util\TimeZone $tz= null) {
+  public function __construct($handle, $fields, TimeZone $tz= null) {
     $this->handle= $handle;
     $this->fields= $fields;
     $this->tz= $tz;
@@ -42,28 +37,40 @@ class ResultSet extends \lang\Object implements Closeable {
   /**
    * Seek to a specified position within the resultset
    *
-   * @param   int offset
+   * @param   int $offset
    * @return  bool success
    * @throws  rdbms.SQLException
    */
-  public function seek($offset) { }
+  public abstract function seek($offset);
 
   /**
    * Iterator function. Returns a rowset if called without parameter,
    * the fields contents if a field is specified or FALSE to indicate
    * no more rows are available.
    *
-   * @param   string field default NULL
+   * @param   string $field default NULL
    * @return  var
    */
-  public function next($field= null) { }
+  public abstract function next($field= null);
 
   /**
    * Close resultset and free result memory
    *
-   * @return  bool success
+   * @return  void
    */
   public function close() { }
+
+  /**
+   * Returns an iterator
+   *
+   * @return php.Iterator
+   */
+  public function getIterator() {
+    if (null === $this->iterator) {
+      $this->iterator= new ResultSetIterator($this);
+    }
+    return $this->iterator;
+  }
 
   /**
    * Returns a string representation of this object