+++ ChangeLog.md
@@ -3,6 +3,6 @@ RDBMS support for the XP Framework: MySQL, Sybase, MSSQL, PostgreSQL, SQLite3, I
 
 ## ?.?.? / ????-??-??
 
+* Added `foreach` support for rdbms.ResultSet (@thekid)
 * Heads up: Changed connections' default charset to UTF-8 (@thekid)
 * Heads up: Converted classes to PHP 5.3 namespaces - (@thekid)
-